+++ demit/apps/__init__.py
@@ -1,0 +1,3 @@
+from demit.apps.palasik import PalasikDemitApp
+
+__all__ = ["PalasikDemitApp"]